+++ native/audio_core/tools/rt/rt_test_support.h
@@ -8,6 +8,7 @@
 namespace rt_test {
 
 void RunPublisherTests();
+void RunBulkRingTests();
 
 }  // namespace rt_test